# api/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Credit Card Learning API
  version: 1.0.0
  description: Simple REST API for bank customer, account, payment, and card domain operations.
servers:
  - url: http://localhost:8081
    description: Customer service
  - url: http://localhost:8082
    description: Account service
paths:
...
  /accounts/{accountId}:
    get:
      summary: Get account by ID
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Account by ID
...
  /payments:
    post:
      summary: Create a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreateRequest'
      responses:
        '201':
          description: Payment created
components:
  schemas:
...
    PaymentCreateRequest:
      type: object
      required: [account_id, payment_amount, payment_method]
      properties:
        account_id:
          type: integer
        payment_amount:
          type: number
        payment_method:
          type: string